fate test update
