+++ examples/benchmark_quality/hetero_linear_regression/fate-linr.py
@@ -28,8 +28,7 @@ from pipeline.utils.tools import load_job_config, JobConfig
 from pipeline.runtime.entity import JobParameters
 
 from federatedml.evaluation.metrics import regression_metric
-from examples.benchmark_quality.parse_result import parse_summary_result
-from fate_test.utils import extract_data
+from fate_test.utils import extract_data, parse_summary_result
 
 
 def main(config="../../config.yaml", param="./linr_config.yaml", namespace=""):
