+++ examples/benchmark_quality/hetero_sbt/fate-sbt.py
@@ -29,8 +29,7 @@ from pipeline.utils.tools import JobConfig
 from pipeline.runtime.entity import JobParameters
 
 from federatedml.evaluation.metrics import regression_metric, classification_metric
-from examples.benchmark_quality.parse_result import parse_summary_result
-from fate_test.utils import extract_data
+from fate_test.utils import extract_data, parse_summary_result
 
 def main(config="../../config.yaml", param="./xgb_config_binary.yaml", namespace=""):
     # obtain config
