+++ python/fate_test/fate_test/flow_test/flow_cli_api.py
@@ -6,7 +6,6 @@ import subprocess
 from pathlib import Path
 
 from prettytable import PrettyTable, ORGMODE
-from fate_flow.entity.types import StatusSet
 from fate_test.flow_test.flow_process import get_dict_from_file
 
 
@@ -60,7 +59,7 @@ class TestModel(object):
                 stdout = json.loads(stdout.decode("utf-8"))
                 if stdout.get('retcode'):
                     self.error_log('job stop: {}'.format(stdout.get('retmsg')) + '\n')
-                if self.query_job() == StatusSet.CANCELED:
+                if self.query_job() == "canceled":
                     return stdout.get('retcode')
             except Exception:
                 return
@@ -174,7 +173,7 @@ class TestModel(object):
             if stdout.get('retcode'):
                 self.error_log('task query: {}'.format(stdout.get('retmsg')) + '\n')
             status = stdout.get("data")[0].get("f_status")
-            if status == StatusSet.SUCCESS:
+            if status == "success":
                 return stdout.get('retcode')
         except Exception:
             return
@@ -477,8 +476,8 @@ class TestModel(object):
         while True:
             time.sleep(5)
             status = self.query_job(job_id=job_id)
-            if status and status in [StatusSet.WAITING, StatusSet.RUNNING, StatusSet.SUCCESS]:
-                if status and status == StatusSet.SUCCESS:
+            if status and status in ["waiting", "running", "success"]:
+                if status and status == "success":
                     return 0
             else:
                 return
@@ -552,7 +551,7 @@ def run_test_api(config_json):
     job.field_names = ['job api name', 'status']
     job.add_row(['job stop', judging_state(test_api.job_api('stop_job'))])
     job.add_row(['job submit', judging_state(test_api.submit_job(stop=False))])
-    job.add_row(['job query', judging_state(False if test_api.query_job() == StatusSet.SUCCESS else True)])
+    job.add_row(['job query', judging_state(False if test_api.query_job() == "success" else True)])
     job.add_row(['job data view', judging_state(test_api.job_api('data_view_query'))])
     job.add_row(['job config', judging_state(test_api.job_config(max_iter=max_iter))])
     job.add_row(['job log', judging_state(test_api.job_api('job_log'))])
