+++ python/fate_test/fate_test/flow_test/flow_sdk_api.py
@@ -6,7 +6,6 @@ from pathlib import Path
 
 from flow_sdk.client import FlowClient
 from prettytable import PrettyTable, ORGMODE
-from fate_flow.entity.types import StatusSet
 from fate_test.flow_test.flow_process import get_dict_from_file
 
 
@@ -71,7 +70,7 @@ class TestModel(object):
                 stdout = self.client.job.stop(job_id=self.job_id)
                 if stdout.get('retcode'):
                     self.error_log('job stop: {}'.format(stdout.get('retmsg')) + '\n')
-                if self.query_job() == StatusSet.CANCELED:
+                if self.query_job() == "canceled":
                     return stdout.get('retcode')
             except Exception:
                 return
@@ -155,7 +154,7 @@ class TestModel(object):
             if stdout.get('retcode'):
                 self.error_log('task query: {}'.format(stdout.get('retmsg')) + '\n')
             status = stdout.get("data")[0].get("f_status")
-            if status == StatusSet.SUCCESS:
+            if status == "success":
                 return stdout.get('retcode')
         except Exception:
             return
@@ -589,8 +588,8 @@ class TestModel(object):
         while True:
             time.sleep(5)
             status = self.query_job(job_id=job_id)
-            if status and status in [StatusSet.WAITING, StatusSet.RUNNING, StatusSet.SUCCESS]:
-                if status and status == StatusSet.SUCCESS:
+            if status and status in ["waiting", "running", "success"]:
+                if status and status == "success":
                     return 0
             else:
                 return
@@ -663,7 +662,7 @@ def run_test_api(config_json):
     job.field_names = ['job api name', 'status']
     job.add_row(['job stop', judging_state(test_api.job_api('stop'))])
     job.add_row(['job submit', judging_state(test_api.submit_job(stop=False))])
-    job.add_row(['job query', judging_state(False if test_api.query_job() == StatusSet.SUCCESS else True)])
+    job.add_row(['job query', judging_state(False if test_api.query_job() == "success" else True)])
     job.add_row(['job view', judging_state(test_api.job_api('view'))])
     job.add_row(['job list', judging_state(test_api.job_api('list/job'))])
     job.add_row(['job config', judging_state(test_api.job_config(max_iter=max_iter))])
